YD-131 Delete user now returns null when the user to be deleted is null
If the 'given:' block fails to create the user, then it cannot be deleted 'in the 'cleanup:' block. The error in the 'cleanup:' block would hide the error of the 'given:' block, making it very hard to find the cause. Now deleteUser simply ignores the call when the given user is null.