YD-196 Messages now include a link when possible
Till now, messages included an embedded User. This implies unnecessary overhead, has the user is available already as part of the buddy data.

The following messages now use a link to the user:
* BuddyConnectRequestMessage
* DisclosureRequestMessage
* DisclosureResponseMessage
* GoalChangeMessage

The following messages embed the user, as the user data might not be available in the buddy data (but now inside _embedded):
* BuddyConnectRequestMessage
* BuddyDisconnectMessage

Additional changes:
* The DisclosureRequestMessage does not embed the target anymore. There is a link, that's sufficient.
* The DisclosureResponseMessage now includes a "related" link for the target message