Changed comment
The comment about the local cache originally said it a local one because of serialization issues, but that's actually not the case. Given that we only use this cache in the analysis engine service, it does not need to be cached across services for as long as we do not scale out the analysis engine service.