YD-227 Time stamp on BuddyAnonymized
Added lastStatusChangeTime to BuddyAnonymized. It is updated when the buddy request is created, accepted, rejected or disconnected.

As this is the first change that requires a schema change, several other changes are included:
* Introduction of updates.yml, the file accumulating includes for all schema changes
* Corrected destination path in liquibase.gradle in task liquibaseDiffChangelog
* Removed the devices table from entity.yml. This shouldn't have been there as the Device class was removed already as part of PR #267. This was missed when merging these changes into the work for the Liquibase introduction.

Unfortunately, we always need to update the generated change log manually due to liquibase/liquibase-hibernate#134. This is included in the wiki procedure: http://wiki.yona.nu/display/DEV/Schema+evolution+with+liquibase