YD-397 Moved password generation to the server
The app can still provide the password upon user sign-up, but this is only temporary. Once both app implementations are updated, this will be disabled.

The Yona password is now returned as part of the private user data. This happens when signing up, but also in other scenarios where the private user data is returned (e.g. GET on user). The app should take the password as returned in the user data and use that in subsequent requests.

Changes:
* Added yonaPassword to the Swagger spec
* Included it in the UserPrivateDto
* Rearranged some crypto-related code
* Made UserController.addUser generate the key if the Yona-Password header was not provided
* Updated the Groovy tests. Most tests now use the server-generated password. Some tests still generated it in the client, which is good for backward compatibility testing.