YD-411 #done Fixed week already full error (#299)
* YD-411 Extended error message
* YD-411 Fixed week already full issue

The cause was that upon a goal change, we create a copy of the goal, but both copies had the same start time. The history item should carry the original creation time, but the "actual" goal should have its creation time updated to the current time. Once the data is corrected, the "week already full" issue will disappear. The fix of this change set will prevent from creating the wrong data again.