YD-419 #done Messages are now automatically processed (#309)
At the time messages are fetched, we now first check whether any messages need to be processed. If so, we do that before returning the messages. As a result, clients never need to trigger processing anymore. This prevents the issue that the first message cannot be interpreted because the second message hasn't been processed yet (the issue causing YD-419).