YD-434 Embed swagger UI
By embedding Swagger UI, we are not longer dependent on the availability of http://petstore.swagger.io, nor do changes there immediately impact us. The Swagger spec was updated to make it work with the current version of Swagger UI. The most notable update is that the parameters are now defined as reusable objects. This makes the spec considerably shorter and it resolves the issue that somehow parameters defined for the path (rather than the operation) weren't recognized anymore.

With this update, it is possible to go to any installation of the Yona server and access Swagger UI on it: http://localhost:8082/swagger-ui/index.html