YD-435 Moved certificate info out of JARs
The following files are not resources in the JAR any longer but taken from the resources folder on the file system:
* appservice/src/main/resources/static/ssl/rootcert.cer
* appservice/src/main/resources/static/vpn/profile.ovpn
* core/src/main/resources/templates/other/apple.mobileconfig.xml

The install scripts now pull these files from GitHub and after that copy node specific files from the specified folder (currently /opt/ope-cloudbees/yona/resources as defined in the Jenkinsfile).

Extended the integration tests to verify the files are properly served.