YD-408 Overlapping activity is merged (#394)
The analysis engine scales each activity up to at least one minute. So this is what could happen:
* User opens app for 20 seconds
* Analysis engine scales this up to 1 minute
* In 20 seconds, the user opens the app again, now for 3 minutes

This should result in an activity of 3 minutes and 40 seconds. In the past, a new activity was created, because start time of the new activity was not equal to the end time of the previous one. Now a new activity is created when the start time is after the end time of the last activity, not when it is equal or before it.