YD-481 Automatically gather Hibernate statistics
The test PerformanceSetupTest is now renamed into HibernateStatsTest and it now automatically gathers Hibernate statistics for various common requests, for the first and the second request. It clears the server caches before performing the requests, to make the influence of the cache clear.

The statistics are stored as a mark down file: appservice/build/reports/tests/intTest/HibernateStatsTest.md. A checked-in copy of the file is maintained as appservice/src/intTest/HibernateStatsTest.md, so we can keep track of changes over time. Copying is a manual action.

The app service is extended with a HibernateStatisticsController that exposes the APIs required by the test. The implementation is done in a HibernateStatisticsService, so we can add a similar controller to for instance the analysisengineservice if we want.

Along with this, the following changes are made:
* Implemented cache clear on UserAnonymizedService
* Added @CacheConfig to ActivityCategoryService, as that was missing
* Introduced a new property enableHibernateStatsAllowed, which controls whether the Hibernate statistics controller is enabled. If it is not enabled, it returns 404 on each operation. It is disabled by default, because it has a (currently unknown) performance overhead. As this could bring trouble to the production environment, it's off by default, but on when the services are started from servers_start.cmd.
* In YonaServer.groovy, renamed toIsoDayString into toIsoDateString and toIsoDateString into toIsoDateTimeString, as the old names were confusing. Added toIsoWeekDateString.
* Extended Goal.groovy with a method to retrieve the goal ID.