YD-145 #done Cache BuddyAnonymized as part of UserAnonymized (#396)
* YD-145 Cache BuddyAnonymized as part of UserAnonymized

The UserAnonymizedDto was already cached inside UserAnonymizedService. Now this DTO holds a set of BuddyAnonymizedDto. Updates to the buddy status now need to be saved as part of the UserAnonymized, so the cache is updated properly.

While making this enhancement, it was noticed that the UserAnonymizedService had an update method that didn't really update the cache. This was corrected and signature of the other update method was simplified.

Next, the update lead to multiple "object references an unsaved transient instance - save the transient instance before flushing" errors. Save calls were added, including a new repository (ActivityRepository).

RemoveUserTest didn't assert the posts to the analysis engine were successful. This was corrected and might need to be done in other tests too.

* YD-145 Updated query statistics

The benefit is very small; the analysis engine service will benefit more, but we're currently not capturing statistics for it.

* YD-145 Addressed review comments

* Renamed variable
* Added explanatory comment to BuddyAnonymized