YD-485 Ignore too large sets of app activities
When a user posts more app activities than the configured threshold, the set of app activities is ignored. This is a workaround for APPDEV-1022: the app accumulates all app activities and posts all historical app activities along with each new one, thus accumulating collections of app activities as large as 745,846 items. By ignoring large sets in the app service, the analysis engine service doesn't get overloaded. Once the issue in the app is fixed, we can remove these provisions again.