YD-492 Guarded DB stats through test
Created a test that guards the database access statistics of the analysis engine service. This service is very performance critical, so it is important that the database access is very limited.

The HibernateStatisticsController is now moved to Core, so each Yona service has the URL path to manage the statistics.