YD-372 Hide applications property on app service
The 'applications' property is of no relevance for the application and thus should not be exposed to it. Removed it from the AppView and updated the tests accordingly.

This breaks backward compatibility, so it can only be merged to the master after the Android app is updated (APPDEV-857) and rolled out. The iOS app is updated and rolled out already.