YD-493 Block all nonmobile numbers
Added a dependency on Google's libphonenumber to verify the numbers. Only numbers of the types MOBILE and FIXED_LINE_OR_MOBILE are accepted. The tests are updated to produce numbers that match Google's definition of mobile numbers.

Extended UserValidationTest with a test that verifies that fixed line numbers are rejected.