YD-492 #done Guarded DB stats through test (#408)
* YD-492 Guarded DB stats through test

Created a test that guards the database access statistics of the analysis engine service. This service is very performance critical, so it is important that the database access is very limited.

The HibernateStatisticsController is now moved to Core, so each Yona service has the URL path to manage the statistics.

* YD-492 Processed review comments

Moved the logic of setEnableStatistics to YonaServer.groovy

* YD-492 Corrected function name

enableStatistics -> setEnableStatistics

* YD-492 Processed review comments

Moved the implementations of all statistics related methods into YonaServer.groovy

* YD-492 Moved common code to Service.groovy

This is how I should have done it immediately. Moved common REST helpers from AppService too.