YD-502 Now use Google Play testing link (#415)
As the app is still in closed beta. Next, corrected the instructions. There used to be a step that said you should first join in the app and then open the URL, but that was wrong. You should after installation immediately open the URL that was provided in the mail.

Once we end closed beta, we should update application.properties and set this property:

yona.email.googlePlayStoreUrl=https://play.google.com/store/apps/details?id=nu.yona.app