YD-495 Add device instance to user
The classic sign-up (both free and invited) now adds a default device to the user. It's also possible to add a device name and operating system when signing up (free sign-up) and when updating the initial data of an invited user.

The Swagger spec has been updated to include the device information in the user's own data. Along with this, the Swagger spec was updated to include the nickname in the base user data, as that's now also available in the user data provided to a buddy.

Device deletion is yet to be handled.