YD-506 Use assertResponseStatus...
It's now used in all applicable places. To make the assertions easier to reuse, they are all moved to the class CommonAssertions.

Along with this fixed a mistake in the test OverwriteUserTest.'Overwrite Richard while being a buddy of Bob and verify Richard does not receive Bob\'s goal conflicts anymore'. That tried to fetch the messages of the overwritten user, but that user does not exist anymore, so it results in a 400. Not sure why this ever passed.