YD-510 Activities are now linked to the device
This applies to app activities and network activities, though the Perl script does not yet support this option, nor do we yet provision device-specific VPN accounts.

A key aspect of this change is that we are communicating the "requesting device ID" in several requests. This allows the app to post app activity on the appropriate URL.

Due to the more extensive use of query parameters, several integration tests failed because of the way we handled query strings. Now YonaServer.groovy handles query strings at the lowest level. Some tests were actually wrong (e.g. used a null URL) but that issue didn't surface yet. With the current changes, some of these issues surfaced and were corrected.

Integration tests for multiple devices can only be written once the device management API is in place.