YD-510 #done Activities are now linked to the device (#425)
* YD-510 Activities are now linked to the device

This applies to app activities and network activities, though the Perl script does not yet support this option, nor do we yet provision device-specific VPN accounts.

A key aspect of this change is that we are communicating the "requesting device ID" in several requests. This allows the app to post app activity on the appropriate URL.

Due to the more extensive use of query parameters, several integration tests failed because of the way we handled query strings. Now YonaServer.groovy handles query strings at the lowest level. Some tests were actually wrong (e.g. used a null URL) but that issue didn't surface yet. With the current changes, some of these issues surfaced and were corrected.

Integration tests for multiple devices can only be written once the device management API is in place.

* YD-510 Deleted AppActivityController.java

It was made empty in the previous commit, but it should have been deleted.

* YD-520 Implemented new device registration

Added it to the existing PR, to allow adding integration tests to it. The new device registration process will soon be described on the wiki (existing page: http://wiki.yona.nu/display/DEV/Flow+-+Add+another+device).

Along with this:
* Added the device API to the Swagger spec, including update and delete, which are yet to be implemented
* Implemented duplicate name checking for device names
* Introduced app version property on DeviceAnonymized

* YD-520 Corrected Swagger spec

The Swagger spec had unsaved changes in Eclipse. Committing these now.

* YD-511 Raw activity data now includes device name

Along with this added missing parts to Swagger spec:
* /users/{userID}/devices/{deviceID}/appActivity/ path
* deviceAppVersion in PostPutUser

Next step is to add integration tests on the raw data to see that activities on different devices are handled properly

* YD-510 Processed review comments

* Renamed DeviceActivityDto into ActivityWithDeviceDto
* Renamed DeviceService.registerNewDevice into addDeviceToUser (to make it align with the already existing method), gave it a UserDeviceDto as parameter and made it cal the already existing method.
* Added basic unit test for new method overload.
* Moved DeviceRegistrationRequestDto into UserDeviceDto and used it for both device registration scenarios: on user sign up and when adding a device
* Moved non-null check from Activity.createInstance into the constructor
* Renamed the parameter deviceId of analyzeAppActivity and buildBaseUrl AnalysisEngineProxyService into deviceAnonymizedId
* Refactored DeviceServiceTest

Along with this:
* Moved NEW_DEVICE_REQUEST_PASSWORD_HEADER from NewDeviceRequestController to Constants

* YD-510 Added integration test for multi-device activity

The test revealed an issue: Activity on one device in a time span within activity on another device is not captured. For this, YD-522 was logged, so it can be solved separately without delaying this PR any further.