YD-497 Updated device management test for VPN
Initially, the test verified deletion of the first (== default) device. Now that the VPN profile moved from the user to the devices, the test started to fail. For backward compatibility, the user still contains the VPN profile, taken from the default device. Deleting the default device implies that the default VPN profile isn't available anymore, causing the test to fail.

The implication of this is that the app first needs to start consuming the VPN profile from the device and then should enable deleting devices.

While running the revised test, another issue came up: The buddy had two devices, even after processing the buddy device change message. The cause of this is that the unprocessed messages were retrieved in descending order (newest first). Due to that, it first processed the delete (which silently failed) and then the add.

Two corrections have been made:
* The messages are now fetched and processed in ascending order (oldest first)
* If the device anonymized cannot be found on the buddy, the delete operation fails.