YD-546 Added device-specific links
Several device-specific links that were already present on the User resource are now added to the Device resource. Once the apps are updated to take the link from the device, the links should be removed from the user (see YD-544).

This includes the following links:
* appleMobileConfig
* postOpenAppEvent
* appActivity
* sslRootCert

The same applies to one property:
* sslRootCertCN

Along with it:
* Made the usage message of HandleDansGuardianLog.pl complete by adding --verbose
* Made the name of the second device in ActivityTest.'Richard works on two devices'() less confusing