YD-561 Fixed "Not avaiable" issue
The first and last name were reported as "Not available, even though the buddy user wasn't migrated yet. This was caused by an off-by-one error on the private data migration version. This fixed now and along with this, message with an embedded user now embed the user with buddy data instead of only public data (i.e. now without first and last name). Finally, the first and last name now fall back to "<nickname>" rather than "Not available", so people get more information in such a case.