KTI_Hazelcast_Breakout (#461)
* Add image base control and liquibase seed version
- Image base allows for non-yona docker images to be
created and used locally for non-jenkins / different
jenkins.
- liquibase seed version allows a different upstream
liquibase seed image to be used because the the liquibase
must fetch material from a published official yona build
which would not exist for local.  A better solution would
be needed if local schema changes needed to be made.

* New imagebase default value, and new defaults for Hazelcast
Also added log level override handling via helm values

* Hazelcast Client setup w/ K8S Discovery
- Add Hazelcast Client library
- Add Hazelcast Kubernetes discovery library
- Add custom properties to indicate hazelcast-client.xml config
as the baked in Spring Boot methods were simply not working
This was based on code provided by Bert, and refactored for hazelcast
client mode

* Helm support for Hazelcast client setup
- hazelcast-client.xml is injected via configmap
- custom property yona.hazelcastConfigFilePath specifies path
to the xml file that is injected.
- Added some other features to enable custom logging levels
via helm values.

* Update K8S helm section to include hazelcast infrastructure
for local testing

* Fixed issue with non-k8s runs

When the Hazelcast config property isn't specified, we instantiate a Hazelcast server rather than a client, thus running with embedded Hazelcast servers again in that scenario.