YD-554 User loading improved considerably
Message loading is still bad and needs improvement

One test fails: nu.yona.server.DeviceManagementTest.'Richard deletes one of his devices; Bob gets the update'

After fixing the failing test, next is to improve message loading.

The current improvement costs a little extra when analyzing the first network traffic for a user. Given that everything is cached for the second and onward, this is not considered an issue.