YD-562 #done Implemented Java locking (#460)
For an unknown reason, the database locking didn't prevent concurrency, so for now resorted to plain Java locking. Given that we currently do not support scale-out, that'll work for now.

To support scale-out, we'll need to make the locking distributed, through Hazelcast. In addition to that, we'll need to add some sort of check again to prevent that we accidentally update and save user information outside the lock. This was done through the UserLockChecker, but that approach doesn't work in combination with a Java lock.