YD-554 #done Improve entity loading (#464)
* YD-554 Base measurement, before improvement

* YD-554 Baseline with SQL statements

The markdown file now includes the SQL statements that were executed

Along with this:
* Corrected property name for Jackson
* Added missing assert to BuddyValidationTest.groovy

* YD-554 Manage log level automatically

When statistics collection is enabled, the log level for the SQL statements is set to enabled. Dropped the separate "isEnabled" flag, as appenders already have a "started" flag.

* YD-554 Load devices anonymized in batch

* YD-554 User loading improved considerably

Message loading is still bad and needs improvement

One test fails: nu.yona.server.DeviceManagementTest.'Richard deletes one of his devices; Bob gets the update'

After fixing the failing test, next is to improve message loading.

The current improvement costs a little extra when analyzing the first network traffic for a user. Given that everything is cached for the second and onward, this is not considered an issue.

* YD-554 Fixed bug in BuddyService.removeDeviceFromBuddy

The removal logic was wrong, causing the association to stay in the database.

* YD-554 Batch loading of activity categories

And corrected batch loading of goals including history items

* YD-554 updated statistics for AnalysisEngineHibernateStatsTest

(downward)

* YD-554 Postpone User DTO creation as much as possible

Inside getMessages. Creating the full user DTO is quite expensive and it used to happen more than once, in different Hibernate sessions.

* YD-554 Improved tryTransferDirectMessagesToAnonymousDestination

Effect isn't visible at this moment, but that could change in the future.

* YD-512 Removed unused method parameters

* YD-554 Addressed review comments

* Removed unnecessary call to assertResponseStatusCreated from BuddyValidationTest
* Used a wired repository for the messageSourceRepository in MessageService rather than the static method on MessageSource