YD-566 @done Shortened most too-long methods (#466)
* YD-566 Shortened most too-long methods

All methods that went beyond 20 lines were shortened, except for a few that would become less readable of shortened (subjective judgment, to be be verified by other committer).

One of the measures is to introduce a class Require that has static methods like isNonNull. Such methods take a value to be asserted and an exception supplier. If the assertion is negative, the exception will be instantiated and thrown. This reduces a construct like

if (a == null) {
    throw ThisAndThatException.missingA();
}

into

Require.isNonNull(a, ThisAndThatException::missingA);

* YD-566 Remove unnecessary if-statement

From the method shouldSendNewGoalConflictMessageForNewConflictingActivity

* YD-566 Processed review comments

* Removed superfluous parameter from UserController.getOwnUser
* Shortened lambda parameter name in ActivityCategoryService.addOrUpdateNewActivityCategories
* Renamed Require.isTrue into Require.that
* Removed Require.isFalse and replaced it with a call to Require.that with a negated condition