YD-570 #done Reload user after each message processing (#472)
The failure scenario is this:
* Richard sends a buddy request to Bob
* Bob accepts it and then changes his personal information
* Richard fetches his messages, which results in automatically processing all to-be-processed MessageService
* The first messages is the buddy acceptance; that succeeds and adds the user anonymized to the buddy
* Given that the user wasn't reloaded after the message was processed, the second one failed as it couldn't find the user anonymized ID on the buddy

The user is now fetched for each message. This will be a little slower when users have multiple messages to be processed, but this is extremely rare and thus not relevant. An integration test is added that reproduced the issue and now guards this scenario.