YD-573 #done Prevent update concurrency (#479)
* YD-573 Prevent update concurrency

The approach of checking and updating for the preparation actions was wrong. The entity was loaded first in multiple threads and then updated within a lock, without reloading it. This is addressed now by consolidating all preparation actions in a single method doPreparationsAndCheckCanAccessPrivateData that is called by the controllers before any service call (as alternative for the old canAccessPrivateData).

Some integration tests had to be updated as processing of messages now also happens earlier, as part of some other requests (not just fetching messages).

* YD-573 Processed review comments

* canAccessPrivateData is now invoked inside the lock, so all entity loading happens inside the lock
* Removed now obsolete tests

* YD-573 Addressed merge conflict

Put the messageService back that was removed with the previous PR.