YD-580 Message action handling now transactional (#481)
Till now, DtoManager.handleAction was not marked with @transactional. Due to that, the services calls from implementations (also in subtypes) were individually starting a transaction, so the handling wasn't atomic. This is fixed now.

Along with this corrected the logging of buddy connect requests. The values "new" and "existing" were interchanged, so existing users where logged as new and vice versa.