YD-584 Add app version code (#491)
This coexists with the app version. The app version is a free-text string that is user readable, the app version code is an integer that is incremented for each and every new build of the app, across releases.

The legacy test for the open app event in DeviceTest.groovy has been updated to also create the user in the legacy way.