Updated VPNSubType in Apple mobile config
Updated based on what's written here: https://docs.openvpn.net/connecting/connecting-to-access-server-with-apple-ios/faq-regarding-openvpn-connect-ios/#Can_I_import_an_OpenVPN_profile_via_an_iOS_mobileconfig_file

Along with this fixed a flipping test: the ActivityCategoriesTest failed a lot. The cause of this is Hazelcast cache propagation. An update on the admin service is not necessarily immediately available in the app service. This propagation is asynchronous. Implemented a polling loop to wait for the propagation.

Other tests might face a similar issue!