YD-591 Add Spring Cloud Sleuth (#505)
That takes care of propagating the B3 headers. It might add extra (unnecessary) spans. We'll see that on the beta test server.

Along with this, the Jenkinsfile now sets the proper namespace to wait for the load test services after deploying to that server. Earlier, it was waiting for the beta services (which were already deployed as part of the previous pipeline stage).