YD-600 Spread index calculation considers DST change
On the date a zone enters or leaves daylight saving time, the day has 23 or 25 hours. This can give issues in the spread calculation. When leaving DST (i.e. on a 25-hour day), the calculated number of minutes for a time stamp after the transition needs to be corrected by subtracting 60 minutes. Right after transitioning into DST, we subtract -60 minutes.