YD-600 Spread index calculation considers DST change (#516)
* YD-600 Spread index calculation considers DST change

On the date a zone enters or leaves daylight saving time, the day has 23 or 25 hours. This can give issues in the spread calculation. When leaving DST (i.e. on a 25-hour day), the calculated number of minutes for a time stamp after the transition needs to be corrected by subtracting 60 minutes. Right after transitioning into DST, we subtract -60 minutes.

* YD-600 Determine correction from timezone offset

Instead of assuming the DST time difference is one hour, it's now being calculated from the offsets.

* YD-600 Processed review comment

Added a null-check for previous transition, in case we ever support time zones that do not have DST.

* YD-600 Processed review comment

Implemented a simpler spread index calculation based on LocalDateTime