YD-601 Correct date for day act. link on goal message
The date was determined by in UTC rather than the user's timezone

Added extra asserts to BasicBuddyTest."Goal conflict of Bob (on iOS and Android) is reported to Richard and Bob". The bug caused the wrong activity (actually an inactivity) to be retrieved, which had a wrong data and a wrong value, but the value wasn't asserted.