YD-601 Further Cinderella bug fixes (#526)
* YD-601 Addressed some Cinderella bugs

* We used to determine the "app last opened date" based on the UTC time. That's now corrected to be the user's timezone. Till now we don't store the user's timezone, but use a constant. This constant moved from UserAnonymizedDto to UserAnonymized.
* Corrected one test case in EditGoalsTest. More tests are to be corrected.

Along with this:
* Made the build repeatable by preventing a timestamp in the build-info.properties
* Made the integration tests log to stdout again

* YD-601 Added comment to mockExistingActivities

Explaining why lenient is being used there.

* YD-601 Corrected and simplified app last opened date

Now the date is always set on the device and the device updates the user entity. The previous commit corrected one occasion of date setting, but missed several others. Now there is just one way to do it.

Along with this, made timeZone in UserAnonymizedDto final.

* YD-601 Correct date for day act. link on goal message

The date was determined by in UTC rather than the user's timezone

Added extra asserts to BasicBuddyTest."Goal conflict of Bob (on iOS and Android) is reported to Richard and Bob". The bug caused the wrong activity (actually an inactivity) to be retrieved, which had a wrong data and a wrong value, but the value wasn't asserted.

* YD-601 Add printErr

The debug log goes to stderr these days. Now it's possible to interject a print statement through printErr

* YD-601 Fixed Cinderella bug

The test is now independent of the moment it's run

* YD-601 Last Cinderella fixes

* Corrected current date calculation in DeviceServiceTest.java
* Corrected calculation of activityRecordedDays in ActivityService.getTotalPageableItems

With this, runs starting at 23:45 and 00:01 complete successfully.

* YD-601 Removed unnecessary TODO comment

* YD-601 Processed review comments

* Renamed GoalConflictMessageDto startDate into activityStartDate
* In DeviceService.setAppLastOpenedDateIfNewer renamed currentDate into today
* In DeviceServiceTest, isolated setting the field in setAppLastOpenedDateField, to prevent spreading the throws clause everywhere

Along with this: fixed bug in UserDevice.setAppLastOpenedDateToNow. It was setting the last opened date to the registration date rather than now.

* YD-601 Triggering PR build