YD-621 Upgrade to Spring Boot 2.1.3
Along with this enable server.use-forward-headers=true again. It was commented out to verify whether our tests now catch that issue.

The real cause of our X-Forwarded-Proto issue (YD-621) was that setting (which apparently used to be default true in Spring 4), not the issue that was fixed through spring-projects/spring-hateoas#758