YD-625 Uncoupled MessageController from specific messages
The MessageController class used to contain specific logic for basically all message types. This implies that creating a new message or considerably changing the implementation of one implies having to change MessageController too. This coupling was too tight.

Instead, we now have the concept of a decorator for messages. A single message instance can be decorated (with embedded entities or links) by multiple decorators that apply to different classes in the type hierarchy of that instance.

Along with this:
* Explicitly asserted the set of links for each and every message type
* Fixed ActivityCommentTest. It often didn't provide the replied message, even though the message being tested was a reply. The old asserts didn't notice that, but the new ones did
* DeviceChangeMessage: added "yona:buddy" link, as already described in the Swagger spec
* DisclosureRequestMessage and DisclosureResponseMessage: added "yona:buddy" link to entity and to the Swagger spec
* BuddyVpnConnectionStatusChangeMessage: Added "yona:buddy" link and added the type to the Swagger spec