YD-637 Disconnect activities before delete device (#553)
* YD-637 Disconnect activities before delete device

Activities have an optional reference to the device on which they occurred. If the user deletes the device, the device reference on the related activities is cleared, but the activities are retained.

Along with this fixed the issue that the cached user anonymized wasn't updated when a device was added or removed.

* YD-637 Assert device activity is retained

Added asserts to verify that though the device is deleted, the activities still exist.