YD-642 Firebase message payload (#557)
* YD-642 Temporarily set a title and body

This way, it is visible that the message reaches the device. Next, we need to agree with the app developers what is exactly needed.

* YD-642 Message content improvements

* Now use a title and message from a translatable resource
* Prepare to take the locale from the device anonymized
* Send the Firebase notification asynchronously, as it can take a significant amount of time
* Expose the messageId on a message DTO, so the app can find the message belonging to a notification

* YD-642 Use the target device locale

Instead of the locale of the device on which the event occurred. Now the solution also works for other messages then goal conflicts.

* YD-642 Use device local for notifications

The device locale is now used for Firebase notifications. The locale is stored on the DeviceAnonymized entity when:
* A new user is registered with an initial device
* A device is added to an existing user
* The open app event is posted with a different locale than before

Along with this:
* Extended Groovy test code with the option to pass headers (next to parameters)
* Explicitly declared the type of the user object (rather than using 'def') on several updated tests