YD-645 Remove last JUnit 4 dependencies (#560)
The class JUnitUtil used org.junit.Assume.assumeThat of JUnit 4. If that assumption failed, JUnit failed the test rather than skipping it. Now the assume is done with org.junit.jupiter.api.Assumptions.assumeTrue.

To prevent such issues, JUnit 4 was removed from the compile-time dependencies of the tests. This required:
* Replacing quite some assertThat calls
* Moving Groovy and Spock from the unit test dependencies and adding them to the integration test dependencies
* Excluding JUnit from the transitive dependencies of the Spring Boot test starter