YD-658 Fixed Eclipse project generation
Needed to include the configuration of the integration tests explicitly.
Along with this:
* Upgraded Gradle from 5.1.1 to 5.6.4

Remaining issue: Eclips complains on the analysisservice project that it is missing two folders. This needs to be resolved manually by removing these two folders from the build path. The cause of this issue is that the analysis service project depends on the test output of the core project (for the repositories), but now Eclipse wants test folders under all source folders of the core project.