YD-658 Various small code improvements (#569)
* YD-658 Optional for BuddyInfoParameters.relatedUserAnonymizedId

* BuddyMessage.BuddyInfoParameters holds a relatedUserAnonymizedId that is sometimes null. This possible absence should be made explicit using {{Optional}}.
* We have an old way of explicitly loading an entity. Nowadays, Hibernate.unproxy exists, so we should adopt that
* In ActivityCommentMessage.getMessagesToBeCascadinglyDeleted, we were using a non-loaded Hibernate proxy. This was masked because the default constructor (required for JPA) of BuddyMessage used the wrong super constructor, thus causing the replies collection to be initialized. This is corrected now.

* YD-658 Various small code improvements

All suggested by SonarQube

* YD-658 Fixed Eclipse project generation

Needed to include the configuration of the integration tests explicitly.
Along with this:
* Upgraded Gradle from 5.1.1 to 5.6.4

Remaining issue: Eclips complains on the analysisservice project that it is missing two folders. This needs to be resolved manually by removing these two folders from the build path. The cause of this issue is that the analysis service project depends on the test output of the core project (for the repositories), but now Eclipse wants test folders under all source folders of the core project.