YD-661 Prevent self-notification (#572)
* YD-660 Fixed detached entity error

This scenario failed:
* User gets invited as buddy
* Invited user overwrites their account
* From the new account, they request the inviting user to become their buddy
* The inviting buddy now cannot process the messages anymore

This was because the direct messages were removed from the named destination and directly inserted in the anonymous destination.

This is now resolved by duplicating the messages. The original one is deleted from the direct destination and a duplicate of it is attached to the anonymous destination.

Along with this:
* Added extra assert to BasicBuddyTest.'Richard finds Bob\'s buddy connect response', to track down a duplicate message issue.
* Updated AnalysisEngineHibernateStatsTest.groovy to lower the number of entities loaded. That somehow improved.
* Removed MessageService.sendDirectMessageAndFlushToDatabase. THe regular sendDirectMessage now always flushes
* FirebaseService.sendMessage now does not save and flush the message anymore but it demands the message to be saved and flushed by its caller.

* YD-661 Prevent self-notification

A copy of activity comment messages "sent" to the sender, as own copy. These now don't cause a Firebase notification anymore.

Note that this commit is based on the change set of YD-660, which should be reviewed and merged first.