YD-653 Made the tests more real-world
The tests were cutting off the mandatory goals, but in real life that never happens. The tests now don't do that anymore.

Updating the tests revealed an issue: The class WeekActivityDto was creating day inacitivty instances for all days of the week, also those that preceded the account creation date. This is fixed now.

It would be good to extend the tests with a global assertion that verifies whether the week activity overview indeed matches the expected values. Now the week activity overview could contain unexpected activities.